docker/python_docker/db_manager/requirements.txt
aiohttp==3.6.2
aiokafka==0.5.2
async-timeout==3.0.1
attrs==19.3.0
cassandra-driver==3.20.2
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
kafka-python==1.4.6
multidict==4.7.3
psycopg2==2.8.4
redis==3.3.11
six==1.13.0
SQLAlchemy==1.3.12
typing-extensions==3.7.4.1
ujson==1.35
yarl==1.4.2
